# .agent/openapi.yaml
openapi: 3.0.3
info:
  title: BuildCRM Backend API
  description: Multi-tenant SaaS platform for construction companies. Manages properties, units, deals, payments, CRM, finance, warehouse, and more.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email, maxLength: 255 }
                password: { type: string, minLength: 8, maxLength: 72 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        full_name: { type: string }
        role: { type: string, enum: [super_admin, company_admin, sales_head, manager, accountant, cashier, foreman, warehouse_manager, broker] }
        can_login: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    AuthResponse:
      type: object
      properties:
        access_token: { type: string }
        access_token_expires_at: { type: string, format: date-time }
        refresh_token: { type: string }
        refresh_token_expires_at: { type: string, format: date-time }
        user:
          $ref: '#/components/schemas/User'
    AuthEnvelope:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/AuthResponse'
...